backend/requirements.txt
# SANGKURIANG Crypto Audit Engine

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Cryptography
cryptography==41.0.7
pynacl==1.5.0
hashlib-compat==1.0.1
pyjwt==2.8.0
bcrypt==4.1.1

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
numpy==1.25.2
pandas==2.1.3

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2

# Security & Audit
bandit==1.7.5
safety==2.3.5
semgrep==1.48.0
bandit==1.7.5

# API & Integration
httpx==0.25.2
aiohttp==3.9.1
celery==5.3.4
dramatiq==1.15.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0
loguru==0.7.2